+++ albums/types.go
@@ -1,0 +1,11 @@
+package albums
+
+type Album struct {
+	ID                    string
+	Title                 string
+	ProductURL            string
+	IsWriteable           bool
+	MediaItemsCount       string
+	CoverPhotoBaseURL     string
+	CoverPhotoMediaItemID string
+}
